codes/requirements.txt
torch==2.6.0
ruamel.yaml==0.17.32
numpy==2.1.1
wandb==0.18.3
nvitop==1.3.2
tensordict==0.7.1
pettingzoo[butterfly]==1.24.3
gymnasium[box2d]==1.0.0
packaging==24.2
dm-meltingpot==2.4.0
git+https://github.com/oxwhirl/smac.git
git+https://github.com/oxwhirl/smacv2.git
